Fix missing ostimer interrupts (#458)
Previous builds with debug enabled we saw cases where after checking
timestamp <= t and before the new MATCH and INTENA was written out
COUNT_VALUE went passed timestamp and timer interrupt stopped firing.

I did some testing and validated even if MATCH value is written but
INTENA=0 the INTRFLAG will not be latched and interrupt will not fire.

To guarantee that the interrupt will fire we must have all the following
completed

MATCH value is latched into the shadow registers
INTENA = 1
Current COUNT_VALUE < timestamp
To ensure these are all correct we first disable timer interrupts and
clear any previously pending interrupt.
Writes to the INTRFLAG are async so we must wait for the WR_RDY bit to
be set before writing to MATCH
Write the new MATCH value
Write INTENA to re-enable timer interrupt
Wait for WR_RDY to indicate the MATCH value was actually latched
Check to make sure the match value we wrote is now after the current
time.
If match value is at or before current time disable interrupts and
return false so the scheduler can run again and set a new alarm value.

I'm not checking for valid timestamp value at the start of the function
since if this value doesn't make any sense or is in the past we will
return false anyways and reset the time value later. This should save a
few clock cycles of execution.