Moved hashcrypt interrupt to embassy interrupt binding (#470)
I am currently working on adding signature verification to
ec-slimloader, where we will be leveraging skboot_authenticate from the
ROM. This will require an alternative binding of the interrupt to a
handler defined in the ROM. Currently embassy-imxrt binds strongly to
this interrupt if the `rt` feature is enabled.

This PR moves the interrupt to the embassy interrupt binding method,
such that the application decides which interrupt to actually bind.

Also moves the interrupt to the toplevel hashcrypt module, instead of
defining it in the more specific hasher module.