Enable UART wakes from deepsleep (#475)
In order to support waking and sending/receiving UART data reliably in
DeepSleep power states, additional registers need to be configured at
the UART layer. If corresponding bits are enabled in the SYSCTL0 HWWAKE
register, peripherals will be allowed to continue clocking and receive
data while the system is in DeepSleep.

---------

Co-authored-by: Colin Donahue <codonah@microsoft.com>