Send + Sync for gpio (#491)
I've been looking at the usage of gpio pins and I noticed a lot of them
are used with mutexes.
However, it seems the mutexes are only used so the pins can be placed in
static memory. This is incredibly wasteful on many fronts.

This PR impls Send and Sync for Flex (which then also causes Input and
Output to be Send and Sync).