Async passthough (#492)
With some refactoring we can be smarter than the compiler (which isn't
too smart when it comes to async).
This saves some flash!

I don't know how much in a *real* application, but all examples combined
save ~4.5kB.

Al behavior and public APIs should still be the same except for the one
in dma/channel.rs, but that one will be an obvious refactor if that
breaks anyone.